--- a/test/rpc-integration/resources/compute-ceil.k
+++ b/test/rpc-integration/resources/compute-ceil.k
@@ -1,0 +1,9 @@
+module COMPUTE-CEIL
+  imports INT
+  imports ID-SYNTAX
+
+  syntax Foo ::= g ( Int )
+
+  rule g ( N:Int ) => N %Int 5
+
+endmodule