~Blocked on https://github.com/runtimeverification/hs-backend-booster/pull/453~

This reverts commit c6edfad1aa4c085e2cc2978c6d7cdc885157b0fb.

This PR brings back the preserves-definiteness attribute addition. ~It sometimes causes rules with question mark variables to apply in Booster, which causes a failure in `pyk`. We cannot re-apply this change until we have a fix for `test-question-mark` variable integration test. We also need to change `test-existential`.~