@goodlyrottenapple 2024-01-12
```
| Test                                     | sam-ceil-fix time | master-bdc5d37 time | (sam-ceil-fix/master-bdc5d37) time
|------------------------------------------|-------------------|---------------------|------------------------------------
| benchmarks/encodepacked-keccak01-spec.k  | 52.14             | 58.74               | 0.8876404494382022
| erc20/ds/transfer-failure-2-b-spec.k     | 59.53             | 65.88               | 0.9036126290224652
| erc20/ds/transferFrom-failure-1-d-spec.k | 83.74             | 75.15               | 1.1143047238855621
| benchmarks/keccak00-spec.k               | 62.35             | 53.58               | 1.1636804777902203
| TOTAL                                    | 257.76            | 253.35000000000002  | 1.0174067495559502
```