@goodlyrottenapple 2024-01-12
```
| Test                                                     | sam-ceil-fix time | master-bdc5d37 time | (sam-ceil-fix/master-bdc5d37) time
|----------------------------------------------------------|-------------------|---------------------|------------------------------------
| BytesTypeTest.test_bytes32(bytes32)                      | 25.99             | 30.75               | 0.8452032520325202
| StoreTest.testStoreLoad()                                | 71.48             | 83.8                | 0.852983293556086
| AllowChangesTest.testAllow()                             | 123.27            | 143.88              | 0.8567556296914095
| ExpectRevertTest.test_ExpectRevert_increasedDepth()      | 151.41            | 174.82              | 0.8660908362887542
| AllowChangesTest.testFailAllowCallsToAddress()           | 122.08            | 140.91              | 0.8663686040735221
| AllowChangesTest.testFailAllowChangesToStorage()         | 123.13            | 141.99              | 0.8671737446299034
| EmitContractTest.testExpectEmitCheckEmitter()            | 78.09             | 89.06               | 0.8768246126207052
| ExpectCallTest.testExpectStaticCall()                    | 78.31             | 88.73               | 0.8825650850895976
| EmitContractTest.testExpectEmitDoNotCheckData()          | 81.19             | 91.55               | 0.8868377935554342
| ExpectCallTest.testExpectRegularCall()                   | 80.24             | 90.14               | 0.8901708453516751
| ExpectRevertTest.testFail_expectRevert_false()           | 96.84             | 107.52              | 0.9006696428571429
| ExpectRevertTest.test_expectRevert_message()             | 94.31             | 103.76              | 0.9089244410177332
| SymbolicStorageTest.testFail_SymbolicStorage1(uint256)   | 144.91            | 159.11              | 0.9107535667148513
| SetUpDeployTest.test_extcodesize()                       | 88.5              | 97.13               | 0.9111500051477401
| ExpectRevertTest.test_expectRevert_true()                | 94.37             | 103.14              | 0.9149699437657554
| ExpectRevertTest.testFail_ExpectRevert_failAndSuccess()  | 97.05             | 106.0               | 0.915566037735849
| ExpectRevertTest.testFail_expectRevert_multipleReverts() | 121.67            | 132.69              | 0.9169492802773381
| AccountParamsTest.test_Nonce_NonExistentAddress()        | 56.88             | 61.76               | 0.9209844559585493
| EmitContractTest.testExpectEmit()                        | 83.2              | 89.92               | 0.9252669039145908
| CounterTest.testSetNumber(uint256)                       | 148.77            | 160.51              | 0.9268581396797708
| CounterTest.testIncrement()                              | 139.86            | 149.49              | 0.9355809753160746
| StoreTest.testGasStoreColdVM()                           | 41.0              | 43.77               | 0.9367146447338359
| PlainPrankTest.test_prank_zeroAddress_true()             | 150.55            | 160.58              | 0.9375389214098891
| PlainPrankTest.test_startPrank_true()                    | 130.42            | 138.82              | 0.9394899870335686
| PlainPrankTest.test_startPrankWithOrigin_true()          | 132.81            | 140.97              | 0.9421153436901468
| AccountParamsTest.testDealSymbolic(uint256)              | 46.44             | 49.21               | 0.9437106279211541
| AccountParamsTest.testEtchConcrete()                     | 68.89             | 72.99               | 0.9438279216331005
| AssertTest.checkFail_assert_false()                      | 34.34             | 36.19               | 0.948880906327715
| PlainPrankTest.test_startPrank_zeroAddress_true()        | 141.3             | 148.15              | 0.9537630779615255
| StoreTest.testStoreLoadNonExistent()                     | 56.02             | 58.67               | 0.9548321118118289
| ExpectRevertTest.test_expectRevert_returnValue()         | 138.94            | 145.43              | 0.9553737193151344
| StoreTest.testGasStoreWarmUp()                           | 65.65             | 68.32               | 0.9609192037470727
| AddrTest.test_builtInAddresses()                         | 25.1              | 26.1                | 0.9616858237547893
| AccountParamsTest.test_GetNonce_true()                   | 41.73             | 43.36               | 0.9624077490774907
| AssumeTest.test_assume_false(uint256,uint256)            | 70.44             | 72.91               | 0.966122616924976
| MethodDisambiguateTest.test_method_call()                | 23.67             | 22.85               | 1.0358862144420131
| BlockParamsTest.testFee(uint256)                         | 40.32             | 38.92               | 1.0359712230215827
| PlainPrankTest.test_stopPrank_notExistent()              | 37.01             | 35.63               | 1.0387314061184394
| AssertTest.test_assert_true_branch(uint256)              | 78.68             | 75.41               | 1.0433629492109802
| BMCLoopsTest.test_countdown_symbolic(uint256)            | 154.75            | 145.23              | 1.0655511946567515
| BlockParamsTest.testCoinBase()                           | 37.02             | 34.11               | 1.085312225153914
| BytesTypeTest.test_bytes4(bytes4)                        | 37.4              | 31.41               | 1.1907035975803884
| TOTAL                                                    | 3654.030000000001 | 3935.6899999999996  | 0.9284344041324397
```